+++ items/services/items_identity/routes/auth/authenticate_password_handler.py
